+++ src/lvmapi/routers/logs.py
@@ -18,6 +18,7 @@ from pydantic import BaseModel, Field
 
 from sdsstools import get_sjd
 
+from lvmapi.routers.notifications import Notification
 from lvmapi.tasks import get_exposure_data_task
 from lvmapi.tools.logs import (
     add_night_log_comment,
@@ -29,7 +30,6 @@ from lvmapi.tools.logs import (
     get_exposures,
     get_night_log_data,
     get_night_log_mjds,
-    get_notifications,
     get_plaintext_night_log,
     get_spectro_mjds,
 )
@@ -104,18 +104,6 @@ class NightLogPostComment(BaseModel):
     ] = None
 
 
-class Notification(BaseModel):
-    """An Overwatcher notification."""
-
-    date: Annotated[datetime, Field(description="The notification datetime.")]
-    message: Annotated[str, Field(description="The notification message.")]
-    level: Annotated[str, Field(description="The level of the notification.")]
-    payload: Annotated[
-        dict | None,
-        Field(description="The payload of the notification."),
-    ] = None
-
-
 router = APIRouter(prefix="/logs", tags=["logs"])
 
 
@@ -174,25 +162,6 @@ async def route_get_exposures(
     return list(map(str, exposures))
 
 
-@router.get("/notifications/{mjd}", summary="Returns notifications for an MJD.")
-async def route_get_notifications(
-    mjd: Annotated[
-        int,
-        Path(description="The SJD for which to list notifications. 0 for current SJD."),
-    ],
-) -> list[Notification]:
-    """Returns a list of notifications for an MJD."""
-
-    mjd = mjd if mjd > 0 else get_sjd("LCO")
-    notifications = await get_notifications(mjd)
-
-    return [
-        Notification(**notification)
-        for notification in notifications
-        if notification["message"] != "I am alive!"
-    ]
-
-
 @router.get("/night-logs", summary="List of night log MJDs.")
 async def route_get_night_logs():
     """Returns a list of MJDs with night log data."""
@@ -246,6 +215,8 @@ async def route_get_night_logs_mjd(
     ],
 ):
     """Returns the night log data for an MJD."""
+
+    from .notifications import route_get_notifications
 
     mjd = mjd if mjd > 0 else get_sjd("LCO")
     data = await get_night_log_data(mjd)